started preparing cluster basis

+++ src/StringCI/Helpers.jl
@@ -3,7 +3,7 @@
 
 Calculates binomial coefficient: n choose k
 """
-function calc_nchk(n::Int,k::Int)
+function calc_nchk(n::Integer,k::Integer)
     accum::Int = 1
     for i in 1:k
         accum = accum * (n-k+i) ÷ i
@@ -17,6 +17,6 @@ end
 
 Looks up binomial coefficient from a precomputed table: n choose k
 """
-@inline function get_nchk(n::Int,k::Int)
+@inline function get_nchk(n,k)
     return binom_coeff[n+1,k+1]
 end
